fix: add better error handling for missing runners
Since we use ephemeral runners, we have a case wherein an exception is
triggered because the runner can't be removed from GitHub because it
(likely) does not exist. To mitigate this, we add logging to show that
this occurred and continue to the next runner. Removing from GitHub
happens automatically over time and doesn't cost the user. Keeping a
running instance because of an exception does, thus with these changes,
the instance will be removed.